renamed to package to match convention cleaned up more imports
